Доработал пути для переименованных файлов

+++ rename1.1.py
@@ -63,23 +63,26 @@ while True:
         root_zip = input('Введите путь места расположения исходного каталога ')
         if os.path.exists(root_zip):
             path_zip = input('Введите путь для распакованных файлов основного архива ')
-            path_temp = os.path.join(path_zip, 'temp')
-            unzip(root_zip, path_zip)
-            unzip(path_zip, path_temp)
-            print('Распаковка файлов завершена')
+            if not os.path.exists(path_zip) or root_zip == path_zip:
+                path_temp = os.path.join(path_zip, 'temp')
+                unzip(root_zip, path_zip)
+                unzip(path_zip, path_temp)
+                print('Распаковка файлов завершена')
 
-            path_xml = os.path.join(path_zip, 'results')
-            if not os.path.exists(path_xml):
-                os.mkdir(path_xml)
-            rename(path_temp)
-            check_and_copy(path_temp, path_xml)
-            delete_files(path_zip)
+                path_xml = os.path.join(path_zip, 'results')
+                if not os.path.exists(path_xml):
+                    os.mkdir(path_xml)
+                rename(path_temp)
+                check_and_copy(path_temp, path_xml)
+                delete_files(path_zip)
 
-            os.chdir(path_zip)
-            shutil.rmtree(path_temp)
-            print('Временные файлы удалены')
+                os.chdir(path_zip)
+                shutil.rmtree(path_temp)
+                print('Временные файлы удалены')
 
-            print('Переименование исходных данных выполнено')
+                print('Переименование исходных данных выполнено')
+            else:
+                print('Вы ввели существующую папку. Это недопустимо. Укажите новую папку или исходную.')
         else:
             print('Вы неправильно указали исходную дирректорию')
     elif choice == 'n':
